challenges/rag-jss-migration-challenge/requirements.txt
functions-framework==3.*
pg8000
cloud-sql-python-connector
Flask
SQLAlchemy
google-cloud-aiplatform>=1.38.0 